add options to file

+++ main.py
@@ -1,21 +1,31 @@
 import csv
+import json
 from datetime import datetime
 import sqlite3
 
 temp_stats = {'abs_max': None, 'abs_min': None, 'abs_avg': None}
 
-db = sqlite3.connect('temp.db')
-db.execute('''
-DROP TABLE IF EXISTS `measures`;
-''')
-db.execute('''CREATE TABLE IF NOT EXISTS measures (
-    meas_time INT PRIMARY KEY,
-    temp NUMERIC,
-    humidity NUMERIC,
-    pressure NUMERIC
-);''')
+options = {
+    'db': True,
+    'out_html': False,
+    'out_raw': True,
+    'in_json': False,
+    'in_csv': True,
+}
 
-db_insert_query = "INSERT INTO measures (meas_time, temp, humidity, pressure) VALUES ('{}','{}','{}','{}');"
+if options['db']:
+    db = sqlite3.connect('temp.db')
+    db.execute('''
+    DROP TABLE IF EXISTS `measures`;
+    ''')
+    db.execute('''CREATE TABLE IF NOT EXISTS measures (
+        meas_time INT PRIMARY KEY,
+        temp NUMERIC,
+        humidity NUMERIC,
+        pressure NUMERIC
+    );''')
+
+    db_insert_query = "INSERT INTO measures (meas_time, temp, humidity, pressure) VALUES ('{}','{}','{}','{}');"
 
 CSV_YEAR = 0
 CSV_MONTH = 1
@@ -45,12 +55,18 @@ def dp(t, h):
 
     return dew_point
 
+if options['out_html']:
+    print '<table>'
+
 
 # that might be a weather station device, not a file
-with open('history_export_2016-12-05T11-55-25.csv', 'rb') as csvfile:
-    weather_station = csv.reader(csvfile, delimiter=';', quotechar='"')
-    # skip header
-    next(weather_station, None)
+with open('history_export_2016-12-05T11-55-25.csv', 'r') as csvfile:
+    if options['in_csv']:
+        weather_station = csv.reader(csvfile, delimiter=';', quotechar='"')
+        # skip header
+        next(weather_station, None)
+    elif options['in_json']:
+        weather_station = json.loads(csvfile.read())['data']
 
     # process data
     for data_row in weather_station:
@@ -71,19 +87,28 @@ with open('history_export_2016-12-05T11-55-25.csv', 'rb') as csvfile:
         measured_day = datetime(int(data_row[CSV_YEAR]), int(data_row[CSV_MONTH]),
                                 int(data_row[CSV_DAY]), int(data_row[CSV_HOUR]), int(data_row[CSV_MINUTE]))
 
-        query = db_insert_query.format(measured_day, data_row[CSV_TEMP], data_row[CSV_RH], data_row[CSV_PRESS])
-        db.execute(query)
-        db.commit()
+        if options['db']:
+            query = db_insert_query.format(measured_day, data_row[CSV_TEMP], data_row[CSV_RH], data_row[CSV_PRESS])
+            db.execute(query)
+            db.commit()
 
         dew_point = dp(float(data_row[CSV_TEMP]), float(data_row[CSV_RH]))
 
-        print "curr: {}, max: {}, min: {}, avg: {}, rh: {}, dp: {}".format(data_row[CSV_TEMP], temp_stats['abs_max'],
+        if options['out_raw']:
+            print "curr: {}, max: {}, min: {}, avg: {}, rh: {}, dp: {}".format(data_row[CSV_TEMP], temp_stats['abs_max'],
                                                                            temp_stats['abs_min'], temp_stats['abs_avg'],
                                                                            data_row[CSV_RH], dew_point)
+        if options['out_html']:
+            print '<tr>'
+            print '<th>curr</th><td>%f</td>' % data_row[TEMP]
+            print '<th>max</th><td>%f</td>' % temp_stats['abs_max']
+            print '<th>min</th><td>%f</td>' % temp_stats['abs_min']
+            print '<th>avg</th><td>%f</td>' % temp_stats['abs_avg']
+            print '<th>rh</th><td>%f</td>' % data_row[CSV_RH] 
+            print '<th>dp</th><td>%f</td>' % dew_point
+            print '</tr>'
 
-
-
-
-        # calculate dew point
-
+if options['out_html']:
+    print '</table>'
+if options['db']:
     db.close()
